Rename some components for clarity
